Disable swap when taking up more than half drive
SLAM will no longer create a swap file, in case the automatically
generated size of the swap file is more than half the total capacity of
the drive, since it doesn't make much sense to use that much of the
total capacity on a swap file. Since it's a file and not a partition the
user can also always just create it themself after. Might implement
feature for adjusting the size of the swap file based on user input in
the future.