Add function for creating a few needed folders
Some of the programs in the base bundles requires some folders to
pre-exist to function properly. This includes things like ZSH that needs
a cache folder that would normally have to be user created to actually
be able to store any command history. This function takes care of
creating that kind of folders.