Use temporary directory for installation files
Prior the script would just copy the files needed for installation in
the chroot over to the root of the new installation. Now it will instead
by utilizing a temporary folder inside of /tmp on the new installation
that will automatically be removed when the script is done to clean up.