Restructure how CSV files are stored
This commit changes how csv files are stored. Instead of having them in
the root dir, they will now have their own designated dir with
subdirectories for different categories. This also includes a change to
parts of the code base, to handle that not all files will be stored in
the same directory, mostly utilizing find instead of ls.