Allow csv files to list depency lists
Bundles can now list other files in the csv format, but with an .dep
file extension. These are designed as a list of dependencies for another
programs, but as they will be installed in the exact same way as the
main bundles, this is mainly to organize the different programs and
their depencies. The reason for the different file extension is to make
sure that the user isn't prompted whether they want to install the
contents of a dependency list, but they instead are only taken in use if
a bundle containing the name with a "D" tag (used to mark that the
following package name should be treated as a local dependency list) is
going to be installed.