Fix problem with copied files not found
Apparently there is some kind of problem with arch-chroot when it comes
to newly created folders, that were created outside the chroot. If you
(before you chroot) create a folder at / on the new installation, they
are there when chrooting, but if you create a folder inside one of the
existing system folders, they aren't there when chrooting. Outside of
the chroot they are visible, but for some reason they aren't inside of
it. The solution for now is simply just to create a new folder at / on
the new installation designated to the SLAM installation.