Updated SLAM to use sfdisk
Prior SLAM would use fdisk, by passing a here-doc to the interactive
point, but it would sometimes stop waiting for the user to confirm the
deletion of prior partitions and thereby stalling the script. This
update moves from using fdisk to using sfdisk without an interactive
prompt, to prevent it from fcking things up.